+++ src/AudioAnalyser.js
@@ -1,54 +1,29 @@
 import React, { useEffect, useState } from "react";
+import {
+  noteFromPitch,
+  noteStrings,
+  centsOffFromPitch
+} from "./util/helpers.js";
 
 const AudioAnalyser = ({ audio }) => {
+  //state variables
   const [pitch, setPitch] = useState();
   const [note, setNote] = useState();
   const [detune, setDetune] = useState();
-
+  //start audio context and analyser
   const audioContext = new (window.AudioContext || window.webkitAudioContext)();
   const analyser = audioContext.createAnalyser();
   const dataArray = new Float32Array(analyser.frequencyBinCount);
-  let source = audioContext.createMediaStreamSource(audio);
+  const source = audioContext.createMediaStreamSource(audio);
   analyser.fftSize = 2048;
   source.connect(analyser);
 
-  //######helper functions to be moved to another file
-  var noteStrings = [
-    "C",
-    "C#",
-    "D",
-    "D#",
-    "E",
-    "F",
-    "F#",
-    "G",
-    "G#",
-    "A",
-    "A#",
-    "B"
-  ];
-
-  function noteFromPitch(frequency) {
-    var noteNum = 12 * (Math.log(frequency / 440) / Math.log(2));
-    return Math.round(noteNum) + 69;
-  }
-
-  function frequencyFromNoteNumber(note) {
-    return 440 * Math.pow(2, (note - 69) / 12);
-  }
-
-  function centsOffFromPitch(frequency, note) {
-    return Math.floor(
-      (1200 * Math.log(frequency / frequencyFromNoteNumber(note))) / Math.log(2)
-    );
-  }
-  //######helper functions to be moved to another file
-
-  var rafID = null; //for animationframe
+  let rafID = null; //request animation frame ID
 
   var MIN_SAMPLES = 0; // will be initialized when AudioContext is created.
   var GOOD_ENOUGH_CORRELATION = 0.9; // this is the "bar" for how close a correlation needs to be
 
+  /// finds and return frequnecy/pitch
   function autoCorrelate(buf, sampleRate) {
     var SIZE = buf.length;
     var MAX_SAMPLES = Math.floor(SIZE / 2);
@@ -75,7 +50,7 @@ const AudioAnalyser = ({ audio }) => {
         correlation += Math.abs(buf[i] - buf[i + offset]);
       }
       correlation = 1 - correlation / MAX_SAMPLES;
-      correlations[offset] = correlation; // store it, for the tweaking we need to do below.
+      correlations[offset] = correlation;
       if (
         correlation > GOOD_ENOUGH_CORRELATION &&
         correlation > lastCorrelation
@@ -86,15 +61,6 @@ const AudioAnalyser = ({ audio }) => {
           best_offset = offset;
         }
       } else if (foundGoodCorrelation) {
-        // short-circuit - we found a good correlation, then a bad one, so we'd just be seeing copies from here.
-        // Now we need to tweak the offset - by interpolating between the values to the left and right of the
-        // best offset, and shifting it a bit.  This is complex, and HACKY in this code (happy to take PRs!) -
-        // we need to do a curve fit on correlations[] around best_offset in order to better determine precise
-        // (anti-aliased) offset.
-
-        // we know best_offset >=1,
-        // since foundGoodCorrelation cannot go to true until the second pass (offset=1), and
-        // we can't drop into this clause until the following pass (else if).
         var shift =
           (correlations[best_offset + 1] - correlations[best_offset - 1]) /
           correlations[best_offset];
@@ -103,12 +69,25 @@ const AudioAnalyser = ({ audio }) => {
       lastCorrelation = correlation;
     }
     if (best_correlation > 0.01) {
-      // console.log("f = " + sampleRate/best_offset + "Hz (rms: " + rms + " confidence: " + best_correlation + ")")
       return sampleRate / best_offset;
     }
     return -1;
-    //	var best_frequency = sampleRate/best_offset;
   }
+
+  let sign = "-";
+  if (detune === 0) {
+    sign = "-";
+  } else if (detune < 0) {
+    sign = "cents ♭";
+  } else if (detune > 0) {
+    sign = "cents ♯";
+  }
+
+  const cleanup = () => {
+    window.cancelAnimationFrame(rafID);
+    analyser.disconnect();
+    source.disconnect();
+  };
 
   function updatePitch() {
     analyser.getFloatTimeDomainData(dataArray);
@@ -122,11 +101,10 @@ const AudioAnalyser = ({ audio }) => {
       // detuneElem.className = "";
       // detuneAmount.innerText = "--";
     } else {
-
       setPitch(Math.round(ac));
       // pitchElem.innerText = Math.round(pitch);
       // var note = noteFromPitch(ac);
-      setNote(noteStrings[noteFromPitch(ac) % 12]);
+      setNote(noteStrings(ac));
       // noteElem.innerHTML = noteStrings[note % 12];
       setDetune(centsOffFromPitch(ac, noteFromPitch(ac)));
       if (detune === 0) {
@@ -140,28 +118,20 @@ const AudioAnalyser = ({ audio }) => {
       }
     }
 
-    if (!window.requestAnimationFrame)
+    if (!window.requestAnimationFrame) {
       window.requestAnimationFrame = window.webkitRequestAnimationFrame;
+    }
     rafID = window.requestAnimationFrame(updatePitch);
-  }
-  let sign = "-";
-  if (detune === 0) {
-    sign = "-";
-  } else if (detune < 0) {
-    sign = "cents ♭";
-  } else if (detune > 0) {
-    sign = "cents ♯";
   }
 
   useEffect(() => {
+    console.log("analyser connected");
+
     updatePitch();
     return () => {
-      source.disconnect();
-      cancelAnimationFrame(rafID);
-      analyser.disconnect();
-      source.disconnect();
+      cleanup();
     };
-  }, []);
+  }, [cleanup]);
 
   return (
     <div>
